Solved problem 437 from leetcode "Path Sum III"
https://leetcode.com/problems/path-sum-iii/description/?envType=study-plan-v2&envId=leetcode-75

+++ src/main/java/org/just/leetcode/study/plan/leetcode75/leetcode435/Solution.java
@@ -33,7 +33,7 @@ public class Solution {
     }
 
     private void pathSum(List<Integer> nodesValue) {
-        int sum = nodesValue.get(nodesValue.size() - 1);
+        long sum = nodesValue.get(nodesValue.size() - 1);
         if (sum == globalTargetSum) {
             globalResult++;
         }
